pyproject.toml
[project]
name = "pricewise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kim Yungju", email = "yjkim101002@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.9",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "langgraph-checkpoint-postgres>=2.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sse-starlette>=3.2.0",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
